[feature] TagsQuery: making TagsQuery based on ExpandableTagsQuery
